% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_ss.R
\name{adonis_ss}
\alias{adonis_ss}
\title{PERMANOVA test for subsampled distance matrices}
\usage{
adonis_ss(
  n1sub,
  n2sub,
  subset_choice = c("Female", "Male", "yes", "no"),
  permutations = 999
)
}
\arguments{
\item{n1sub}{First subsample distance matrix.}

\item{n2sub}{Second subsample distance matrix.}

\item{subset_choice}{Character. One of \code{"Female"}, \code{"Male"}, \code{"yes"} or \code{"no"}.}

\item{permutations}{Number of permutations passed to \code{vegan::adonis2()}.}
}
\value{
First row of the \code{adonis2()} table, including a \code{p.value} column.
}
\description{
PERMANOVA test for subsampled distance matrices
}
\keyword{internal}
