% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_effect.R
\name{prob_effect}
\alias{prob_effect}
\title{Estimate probability of effect between two similarity structures}
\usage{
prob_effect(
  fitting_A = NULL,
  fitting_B = NULL,
  smooth_A = NULL,
  smooth_B = NULL
)
}
\arguments{
\item{fitting_A}{Fitting output for group A.}

\item{fitting_B}{Fitting output for group B.}

\item{smooth_A}{Smoothing output for group A.}

\item{smooth_B}{Smoothing output for group B.}
}
\value{
A probability or a list of probabilities comparing A and B.
}
\description{
Estimate probability of effect between two similarity structures
}
