% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rskew_fixed.R
\name{rskew_fixed}
\alias{rskew_fixed}
\title{Generate skew-normal random values with fixed mean and variance}
\usage{
rskew_fixed(n, alpha, mu_target = 0, var_target = 1)
}
\arguments{
\item{n}{Number of observations.}

\item{alpha}{Shape parameter of the skew-normal distribution.}

\item{mu_target}{Target mean.}

\item{var_target}{Target variance.}
}
\value{
Numeric vector of simulated values.
}
\description{
Generate skew-normal random values with fixed mean and variance
}
\keyword{internal}
