% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_structure.R
\name{similarity_structure}
\alias{similarity_structure}
\title{Estimate similarity structure between two samples}
\usage{
similarity_structure(
  n1,
  n2,
  N_init,
  N_fin,
  num_N,
  num_repet,
  replace = TRUE,
  test,
  test_args = NULL,
  alpha,
  seed = TRUE,
  sav = FALSE,
  sav_file = NULL,
  plotting = TRUE,
  title_plot = NULL
)
}
\arguments{
\item{n1}{First sample. It can be a vector or a matrix.}

\item{n2}{Second sample. It can be a vector or a matrix.}

\item{N_init}{Initial subsample size.}

\item{N_fin}{Final subsample size.}

\item{num_N}{Number of equidistant subsample sizes to evaluate.}

\item{num_repet}{Number of repetitions for each subsample size.}

\item{replace}{Logical. Should sampling be done with replacement?}

\item{test}{Statistical test to use. Either \code{"t-test"}, \code{"ks-test"}, or a custom function.}

\item{test_args}{Optional list of additional arguments passed to the test.}

\item{alpha}{Significance level.}

\item{seed}{Seed used by the parallel backend.}

\item{sav}{Logical. Should the empirical PN-alpha object be saved?}

\item{sav_file}{Optional file path for saving the empirical PN-alpha object.}

\item{plotting}{Logical. Should the fitted and smoothed curves be plotted?}

\item{title_plot}{Optional title for the plot.}
}
\value{
A list with two components:
\describe{
\item{empiric_PNalpha}{Empirical P(N | >alpha) curve.}
\item{simil_struct_Ns}{Estimated similarity structure.}
}
}
\description{
Estimate similarity structure between two samples
}
