% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_testing.R
\name{subsampling_testing}
\alias{subsampling_testing}
\title{Apply a statistical test to two subsamples}
\usage{
subsampling_testing(
  n1,
  n2,
  nsub,
  test = "t-test",
  test_args = NULL,
  replace = TRUE
)
}
\arguments{
\item{n1}{First sample. It must be either a vector or a matrix.}

\item{n2}{Second sample. It must be of the same type as \code{n1}.}

\item{nsub}{Integer. Size of the subsamples to draw from \code{n1} and \code{n2}.}

\item{test}{Statistical test to use. Either \code{"t-test"}, \code{"ks-test"}, or a
custom function. Custom functions must return an object containing a
\code{p.value} element.}

\item{test_args}{Optional list. Additional arguments passed to the test
function.}

\item{replace}{Logical. Should subsampling be performed with replacement?}
}
\value{
A numeric value corresponding to the p-value of the selected test.
}
\description{
Draws a subsample of size \code{nsub} from each of two samples and applies a
statistical test to the resulting subsamples. The function supports vectors
and matrices. If matrices are provided, rows are sampled.
}
\keyword{internal}
